Support Landlock ABI v3 (file truncation)
Kernel 6.2 will make it possible to restrict file truncation using
Landlock. Before kernel 6.2, truncate(2) and ftruncate(2) system calls
always worked, even when in a Landlock-confined process or using a
Landlock-confined file descriptor.

File truncation support in Landlock works very similarly to "file
write" support, as it is also a way to modify file contents. Most
users will want to always use the "truncate" right hand in hand with
the "write file" right.

Users of landlock.V1 and landlock.V2 are advised to update their
RestrictPath() invocations to use landlock.V3. When used together with
ROFiles, RWFiles, RODirs and RWDirs, the changes will be backwards
compatible, provided that all file hierarchies which will be truncated
have been spelled out in the RestrictPath() invocation.

Programs that do not use truncate(2) or ftruncate(3) can upgrade to
landlock.V3 without compatibility issues.

If you are passing opened file descriptors between Landlocked and
non-Landlocked processes, please refer to the kernel Landlock
documentation for details on the truncate semantics in this case.

This patch was tested with kernel 6.2-rc3.