Rename DialTCP to ConnectTCP
The name "DialTCP" is misleading, because it is actually the
connect(2) syscall which is being restricted.  This makes a difference
in the scenario where a future version of Landlock makes socket(2)
restrictable.  (net.Dial() creates sockets as well, so it would be
confusing if permitting landlock.DialTCP() was not enough to
net.Dial() a TCP socket) -- It's better to keep the terminology
aligned with kernel concepts.

I am shamelessly assuming that noone has picked up the DialTCP changes
from yesterday yet.  (If you did, my apologies.)  Better to have a
brief window of incompatibility than to carry out long API migrations
based on speculative API usages.