syscall: Rename NetServiceAttr to NetPortAttr
This was originally named in line with an earlier, unstable version of
the networking support, and I missed to update it.  We do not give
backwards compatibility guarantees on the syscall module, but I
was unable to find direct dependencies on it on the internet.

If this breaks you, please reach out with a Github issue.