Work around libpsx accessing /proc
Libpsx with CGO lists the directory `/proc/$PID/task` just after
Landlock enforcement on each OS thread.  This is problematic, because
commonly, Landlock policies will forbid this file system operation,
and libpsx responds to the `EACCES` by crashing the process.

This workaround silently allow-lists the `LANDLOCK_ACCESS_FS_READ_DIR`
right (and only that!) on the `/proc/$PID/task` directory, so that
existing Landlock enforcements, at least within the same Go process,
continue to work as expected.

This workaround will fix Go-Landlock with CGO for most programs, but
unfortunately a corner case stays broken: when the Go process executes
another libpsx-using program.  Users who need to use libpsx in a
subprocess are advised to either:

* depend on libpsx 1.2.71 in their go.mod file and don't upgrade
* compile their program without CGO

The proper long-term solution for this problem is to add support for
multithreaded Landlock enforcement to the kernel in the form of a
feature for the
[*landlock_restrict_self*(2)](https://man.gnoack.org/2/landlock_restrict_self)
system call.

Fixes: #39
Link: https://wiki.gnoack.org/LandlockMultithreadedEnforcement
Link: https://bugzilla.kernel.org/show_bug.cgi?id=220644
Link: https://github.com/landlock-lsm/go-landlock/issues/39
Suggested-by: Jann Horn